hand/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stservo"
version = "1.0.0"
authors = [
    {name = "iltlo", email = "iltlo@connect.hku.hk"},
]
description = "Python library for controlling STServos with GUI interface"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["servo", "stservo", "robotics", "hardware", "control"]
dependencies = [
    "pyserial>=3.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gui = ["tkinter"]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/iltlo/waveshare_stservo_python"
Repository = "https://github.com/iltlo/waveshare_stservo_python"
Documentation = "https://github.com/iltlo/waveshare_stservo_python/blob/main/README.md"
Issues = "https://github.com/iltlo/waveshare_stservo_python/issues"

[project.scripts]
stservo-gui = "stservo.gui.servo_gui:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"stservo" = ["../config/*.yaml"]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
